add separate algorithms.rs file to model module
